# packages/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/boards": {
      "get": {
        "operationId": "BoardController_getBoards",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardsGetResponse"
                }
              }
            }
          },
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardsGetResponse"
                }
              }
            }
          }
        },
        "summary": "Get all boards",
        "tags": ["Boards"]
      },
      "post": {
        "operationId": "BoardController_createBoard",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BoardCreateRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardGetOneResponse"
                }
              }
            }
          },
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardGetOneResponse"
                }
              }
            }
          }
        },
        "summary": "Create a new board",
        "tags": ["Boards"]
      }
    },
    "/boards/{boardId}": {
      "get": {
        "operationId": "BoardController_getBoard",
        "parameters": [
          {
            "name": "boardId",
            "required": true,
            "in": "path",
            "description": "The ID of the board to retrieve",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardGetOneResponse"
                }
              }
            }
          },
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardGetOneResponse"
                }
              }
            }
          }
        },
        "summary": "Get a board by ID",
        "tags": ["Boards"]
      },
      "put": {
        "operationId": "BoardController_updateBoard",
        "parameters": [
          {
            "name": "boardId",
            "required": true,
            "in": "path",
            "description": "The ID of the board to update",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BoardPutRequestDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          },
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardGetOneResponse"
                }
              }
            }
          }
        },
        "summary": "Update a board by ID",
        "tags": ["Boards"]
      },
      "delete": {
        "operationId": "BoardController_deleteBoard",
        "parameters": [
          {
            "name": "boardId",
            "required": true,
            "in": "path",
            "description": "The ID of the board to delete",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Delete a board by ID",
        "tags": ["Boards"]
      }
    },
    "/boards/{boardId}/sync": {
      "post": {
        "operationId": "BoardController_syncBoard",
        "parameters": [
          {
            "name": "boardId",
            "required": true,
            "in": "path",
            "description": "The ID of the board to sync",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Sync a board by ID",
        "tags": ["Boards"]
      }
    },
    "/posts/search": {
      "post": {
        "operationId": "PostController_searchPosts",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostsSearchRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Returns paginated posts with cursor for next page",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostsGetResponse"
                }
              }
            }
          },
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostsGetResponse"
                }
              }
            }
          }
        },
        "summary": "Search for posts across all boards with cursor pagination",
        "tags": ["posts"]
      }
    },
    "/posts/{postId}": {
      "get": {
        "operationId": "PostController_getPost",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "description": "The ID of the post to get",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostGetResponse"
                }
              }
            }
          }
        },
        "summary": "Get a post by ID",
        "tags": ["posts"]
      }
    },
    "/posts/{postId}/sync": {
      "post": {
        "operationId": "PostController_syncPost",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "description": "The ID of the post to sync",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Sync a post by ID",
        "tags": ["posts"]
      }
    },
    "/posts/{postId}/apply-decision": {
      "post": {
        "operationId": "PostController_applyDecision",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "description": "The ID of the post to apply the decision to",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyDecisionRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Apply a decision to a post",
        "tags": ["posts"]
      }
    },
    "/tags/{tagId}/generate-description": {
      "post": {
        "operationId": "TagController_generateDescription",
        "parameters": [
          {
            "name": "tagId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagGenerateDescriptionResponseDto"
                }
              }
            }
          }
        },
        "tags": ["Tags"]
      }
    },
    "/boards/{boardId}/tags": {
      "get": {
        "operationId": "TagController_getTags",
        "parameters": [
          {
            "name": "boardId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagsGetResponse"
                }
              }
            }
          }
        },
        "tags": ["Tags"]
      },
      "put": {
        "operationId": "TagController_putTags",
        "parameters": [
          {
            "name": "boardId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TagsPutRequestDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": ["Tags"]
      }
    },
    "/boards/{boardId}/context": {
      "get": {
        "operationId": "BoardContextController_getBoardContext",
        "parameters": [
          {
            "name": "boardId",
            "required": true,
            "in": "path",
            "description": "Board ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardContextGetDto"
                }
              }
            }
          }
        },
        "summary": "Get board context",
        "tags": ["Board Context"]
      },
      "put": {
        "operationId": "BoardContextController_updateBoardContext",
        "parameters": [
          {
            "name": "boardId",
            "required": true,
            "in": "path",
            "description": "Board ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BoardContextPutRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardContext"
                }
              }
            }
          }
        },
        "summary": "Update board context",
        "tags": ["Board Context"]
      }
    },
...
  },
  "info": {
    "title": "BoardBot API",
    "description": "BoardBot API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "BoardGet": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "vendor": {
            "type": "string",
            "nullable": true,
            "enum": ["fider"]
          },
          "autoTriggerModeration": {
            "type": "boolean"
          },
          "autoApplyDecision": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "vendor",
          "autoTriggerModeration",
          "autoApplyDecision",
          "createdAt",
          "updatedAt"
        ]
      },
      "BoardsGetResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BoardGet"
            }
          }
        },
        "required": ["data"]
      },
      "BoardGetOneResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/BoardGet"
          }
        },
        "required": ["data"]
      },
      "BoardPutRequestDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 2
          },
          "description": {
            "type": "string"
          },
          "autoTriggerModeration": {
            "type": "boolean"
          },
          "autoApplyDecision": {
            "type": "boolean"
          }
        },
        "required": ["title", "description"]
      },
      "BoardCreateRequestDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "description": "The name of the board",
            "example": "My Board"
          }
        },
        "required": ["name"]
      },
      "PostProcessingStatusEnum": {
        "type": "string",
        "enum": ["pending", "awaiting_manual_review", "completed", "failed"],
        "description": "Filter posts by processing statuses"
      },
      "PostsSearchRequestDto": {
        "type": "object",
        "properties": {
          "statuses": {
            "description": "Filter posts by processing statuses",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PostProcessingStatusEnum"
            }
          },
          "cursor": {
            "type": "string",
            "description": "Cursor for pagination (typically the ID of the last post from previous page)"
          },
          "limit": {
            "type": "number",
            "description": "Maximum number of posts to return",
            "default": 20,
            "minimum": 1,
            "maximum": 100
          },
          "boardIds": {
            "description": "Filter posts by board IDs",
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        }
      },
      "PostStatusEnum": {
        "type": "string",
        "enum": [
          "open",
          "planned",
          "started",
          "completed",
          "declined",
          "duplicate"
        ]
      },
      "TagGet": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "externalId": {
            "type": "string"
          }
        },
        "required": ["id", "title", "description", "externalId"]
      },
      "ModerationDecisionEnum": {
        "type": "string",
        "enum": ["accepted", "rejected", "unknown"],
        "description": "The decision of the moderation"
      },
      "ModerationReasonEnum": {
        "type": "string",
        "enum": [
          "multiple_suggestions",
          "is_a_question",
          "is_spam_or_inappropriate",
          "is_advertisement",
          "is_bug_report",
          "is_not_understandable"
        ],
        "description": "The reason of the moderation"
      },
      "ModerationDecision": {
        "type": "object",
        "properties": {
          "decision": {
            "description": "The decision of the moderation",
            "allOf": [
              {
                "$ref": "#/components/schemas/ModerationDecisionEnum"
              }
            ]
          },
          "reason": {
            "description": "The reason of the moderation",
            "allOf": [
              {
                "$ref": "#/components/schemas/ModerationReasonEnum"
              }
            ]
          },
          "reasoning": {
            "type": "string",
            "description": "The AI reasoning"
          }
        },
        "required": ["decision"]
      },
      "DuplicatePostsDecisionStatusEnum": {
        "type": "string",
        "enum": ["success", "failed"],
        "description": "The status of the duplicate posts decision"
      },
      "DuplicatePostsDecisionEnum": {
        "type": "string",
        "enum": ["duplicate", "not_duplicate", "unknown"],
        "description": "The decision of the duplicate posts"
      },
...
      "DuplicatePostsDecision": {
        "type": "object",
        "properties": {
          "status": {
            "description": "The status of the duplicate posts decision",
            "allOf": [
              {
                "$ref": "#/components/schemas/DuplicatePostsDecisionStatusEnum"
              }
            ]
          },
          "decision": {
            "description": "The decision of the duplicate posts",
            "allOf": [
              {
                "$ref": "#/components/schemas/DuplicatePostsDecisionEnum"
              }
            ]
          },
          "duplicatePosts": {
            "description": "List of duplicate posts with their reasoning",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DuplicatePost"
            }
          },
          "reasoning": {
            "type": "string",
            "description": "The AI reasoning"
          }
        },
        "required": ["status", "decision", "duplicatePosts"]
      },
      "TagAssignmentDecisionStatusEnum": {
        "type": "string",
        "enum": ["success", "failed"],
        "description": "The status of the tag assignment"
      },
      "TagAssignmentDecision": {
        "type": "object",
        "properties": {
          "status": {
            "description": "The status of the tag assignment",
            "allOf": [
              {
                "$ref": "#/components/schemas/TagAssignmentDecisionStatusEnum"
              }
            ]
          },
          "tagIds": {
            "description": "The tag IDs",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "reasoning": {
            "type": "string",
            "description": "The AI reasoning"
          }
        },
        "required": ["status", "tagIds"]
      },
      "PostDecision": {
        "type": "object",
        "properties": {
          "moderation": {
            "$ref": "#/components/schemas/ModerationDecision"
          },
          "duplicatePosts": {
            "$ref": "#/components/schemas/DuplicatePostsDecision"
          },
          "tagAssignment": {
            "$ref": "#/components/schemas/TagAssignmentDecision"
          }
        }
      },
      "PostGet": {
        "type": "object",
        "properties": {
          "processingStatus": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PostProcessingStatusEnum"
              }
            ]
          },
          "postUrl": {
            "type": "string"
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PostStatusEnum"
              }
            ]
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagGet"
            }
          },
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "decision": {
            "$ref": "#/components/schemas/PostDecision"
          },
          "board": {
            "$ref": "#/components/schemas/BoardGet"
          },
          "postCreatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "externalId": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "processingStatus",
          "id",
          "title",
          "description",
          "board",
          "postCreatedAt",
          "externalId",
          "createdAt",
          "updatedAt"
        ]
      },
      "PostsGetResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PostGet"
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true
          }
        },
        "required": ["data", "nextCursor"]
      },
      "IncludesGetDto": {
        "type": "object",
        "properties": {
          "posts": {
            "description": "The related posts",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PostGet"
            }
          }
        },
        "required": ["posts"]
      },
      "PostGetResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PostGet"
          },
          "includes": {
            "$ref": "#/components/schemas/IncludesGetDto"
          }
        },
        "required": ["data", "includes"]
      },
      "ApplyModerationDecision": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "enum": [
              "multiple_suggestions",
              "is_a_question",
              "is_spam_or_inappropriate",
              "is_advertisement",
              "is_bug_report",
              "is_not_understandable"
            ]
          }
        }
      },
      "ApplyDuplicatePostsDecision": {
        "type": "object",
        "properties": {
          "duplicatePostId": {
            "type": "string",
            "description": "The ID of the applied duplicate post"
          }
        }
      },
      "ApplyTagAssignmentDecision": {
        "type": "object",
        "properties": {
          "tagIds": {
            "description": "The IDs of the applied tags to the post",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "ApplyDecisionRequestDto": {
        "type": "object",
        "properties": {
          "moderation": {
            "$ref": "#/components/schemas/ApplyModerationDecision"
          },
          "duplicatePosts": {
            "$ref": "#/components/schemas/ApplyDuplicatePostsDecision"
          },
          "tagAssignment": {
            "$ref": "#/components/schemas/ApplyTagAssignmentDecision"
          }
        }
      },
      "TagGenerateDescriptionResponseDto": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string"
          }
        },
        "required": ["description"]
      },
      "TagsGetResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagGet"
            }
          }
        },
        "required": ["data"]
      },
      "TagPut": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        },
        "required": ["id", "description"]
      },
      "TagsPutRequestDto": {
        "type": "object",
        "properties": {
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagPut"
            }
          }
        },
        "required": ["tags"]
      },
      "BoardContextGetDto": {
        "type": "object",
        "properties": {
          "productDescription": {
            "type": "string"
          },
          "productGoals": {
            "type": "string"
          }
        },
        "required": ["productDescription", "productGoals"]
      },
      "BoardContextPutRequestDto": {
        "type": "object",
        "properties": {
          "productDescription": {
            "type": "string"
          },
          "productGoals": {
            "type": "string"
          }
        }
      },
      "BoardContext": {
        "type": "object",
        "properties": {}
      },
...
    }
  },
  "security": [
    {
      "bearer": []
    }
  ]
}