# packages/backend/src/modules/board/controllers/board-context.controller.ts
import { Body, Controller, Get, Param, Put } from "@nestjs/common"
import { ApiOperation, ApiParam, ApiTags } from "@nestjs/swagger"
import { BoardContextPutRequestDto } from "src/modules/board/models/dto/board-context-put.request.dto"
import { BoardContextService } from "../services/board-context.service"
import { toBoardContextGetDto } from "src/modules/board/models/dto/board-context-get.dto"
import { BoardAccess } from "src/modules/user/decorators/board-access.decorator"

@ApiTags("Board Context")
@Controller("boards/:boardId/context")
@BoardAccess({ boardIdParam: "boardId" })
export class BoardContextController {
  constructor(private readonly boardContextService: BoardContextService) {}

  @Get()
  @ApiOperation({ summary: "Get board context" })
  @ApiParam({ name: "boardId", type: String, description: "Board ID" })
  async getBoardContext(@Param("boardId") boardId: string) {
    const context = await this.boardContextService.getBoardContext(boardId)
    return toBoardContextGetDto(context)
  }

  @Put()
  @ApiOperation({ summary: "Update board context" })
  @ApiParam({ name: "boardId", type: String, description: "Board ID" })
  async updateBoardContext(
    @Param("boardId") boardId: string,
    @Body() data: BoardContextPutRequestDto,
  ) {
    return this.boardContextService.updateBoardContext(boardId, data)
  }
}